Add error handling if unauthenticated user try to delete user

+++ src/main/java/com/teamc/bioskop/Service/impl/UserServiceImplements.java
@@ -127,7 +127,9 @@ public class UserServiceImplements implements UserService, UserDetailsService {
      */
     @Override
     public void deleteUserById(Authentication authentication) {
-
+        if (authentication == null){
+            throw new ResourceNotFoundException("Not authenticated, forbidden");
+        }
         String userAuth = authentication.getName();
         User user = this.userRepository.findByUsername(userAuth);
         this.userRepository.delete(user);
